Removes droppable helper text on device drop
The droppable area displays a helper text that states "Drop Devices Here", which now now detached when a device is dropped into the area.

The reason it is being detached, rather than deleted, is because once the ability to remove devices from "Your Devices" is implemented, the helper text will be re-appended when the area is empty again.