Added ability to search for devices from API
The search_devices() function takes category names and queries the search api for the first one that is also a category. So far this has been returning the correct device wiki object for the category. The title and image thumbnail from the object can then be used by the web app.

Mootools was added to the .gitignore