Added limit variable, and fixed next and prev functions
The limit variable will be used to increase and decrease the offset each time a user clicks the "next" or "prev" buttons. The innerHTML of the device container is now cleared each next or prev call as well.

For requests to the search api that return 0 hits, a console message is printed that notifies that there was no result found for the category.