Copied and adjusted HTML and CSS prototypes
Index.html and Style.css are adjusted versions of the early graphical prototypes I wrote. The drag and drop is no longer functional, because it needs to be rewritten with mootools. That will be coming shortly. For now, index.html is serving as a test page to demonstrate API access. This is also the general style I have planned for the final product.