Added fixed height/width dimensions to device thumbnails
This helps with choppiness in displaying device blocks. When the dimensions weren't set, the blocks would jump around if the images weren't completely loaded. Now, the device block will still load if images aren't finished loading, but the blocks don't jump around when adjusting for the thumbnails dimension.

Since all thumbnails are 96 x 72, so this shouldn't cause any issue with resizing.