Moved script tags in index.html, changed shadow colors
The script tags are now at the bottom of the body of the index.html, and no longer wait for the 'domready' event to begin. Hopefully this helps with the initial loading of the page.

The box-shadow on the device blocks has been changed to a light grey and black when hovered over. The devices container now expands to fit all device blocks even when page dimensions shift.