Prototype drag and drop
With help from the Mootools Drag.Cart demo, the gear bag now has experimental drag and drop. So far I haven't been able to check whether or not the gear bag contains an element, so duplicates can be added. Also, whether or not elements can be removed from the gear bag still needs to considered.