Cancel device drop if already in gear bag
Added a new function in_drop() which checks the elements already in the gear bag with the new device that is currently being dropped in the gear bag. If the function returns false, the device being dropped is then injected into the gear bag. If the function returns true, the device clone that is being dragged is destroyed, as opposed to being injected into the gear bag.