Help text and other DB-related bug fixes
The "Drop Devices Here" help text was being destroyed when there were no devices in the database to load. This destroyed the purpose of the help text, which is to show users what the point of the droppable area on the right is. Exceptions were also being thrown when trying to access result sets that had null values.

The api requests were getting permission errors, so by adding 'callback=?' to the request URLs, the problems were solved.