Fixed bug that doesn't show helper text
When the database has been created, but does not have any devices saved in it, the gear bag helper text is overwritten with an empty string.

There is now a condition in place that checks if the length of the result.rows array is greater than 0 before replacing the innerHTML of the drop_device element.