Fixed next/prev bug where multiple presses loads too many devices
Whenever prev or false is clicked, a lock boolean is set to true, which prevents any new calls to prev or false from running. After all requests have been recieved and added to the devices container, the lock is lifted and next/prev will be able to load the next page.

Before the lock, if next/prev was clicked more than once, it would load |limit| * (number of times next/prev was clicked) devices at once.